+++ apps/backoffice/urls.py
@@ -4,6 +4,8 @@ from . import views
 app_name = 'backoffice'
 
 urlpatterns = [
+    path('dashboard/', views.dashboard, name='dashboard'),
+
     path('cotacao/importar/', views.cotacao_import, name='cotacao_import'),
     path('cotacao/modelo/', views.download_modelo_cotacao, name='download_modelo'), # 🚀 Rota do download